+++ client/src/pages/index/components/SearchBar/withDropdown.jsx
@@ -59,7 +59,7 @@ class SearchBar extends Component {
         overlay={menu}
       >
         <Search
-          placeholder="歌曲/专辑/艺人"
+          placeholder="歌曲 | 专辑 | 艺人"
           defaultValue={keyword || ''}
           onSearch={this.onSearch}
           enterButton